requirements.txt
Werkzeug==2.0.1
Flask==2.0.1
easydict==1.9
flask_expects_json==1.5.0
PyYAML==5.4.1
